reorganized components directory
