+++ frontend/foodieland-app/components/categories/CategoriesSection.tsx
