+++ frontend/foodieland-app/components/featured/FeaturedSection.tsx
