+++ frontend/foodieland-app/components/featured/FeaturedSlider.tsx
