+++ frontend/foodieland-app/components/index.ts
@@ -1,42 +1,42 @@
-import Title from "@/components/ui/Title";
+import Title from "@/components/shared/Title";
 import Logo from "@/components/shared/Logo";
 import Button from "@/components/ui/Button";
 import Inbox from "@/components/shared/Inbox";
-import Footer from "@/components/shared/Footer";
+import Footer from "@/components/layout/Footer";
 import LikeButton from "@/components/ui/LikeButton";
 import LinkButton from "@/components/ui/LinkButton";
 import InputGroup from "@/components/ui/InputGroup";
 import UserInfo from "@/components/shared/UserInfo";
 import Container from "@/components/shared/Container";
-import Description from "@/components/ui/Description";
+import Description from "@/components/shared/Description";
 import CheckButton from "@/components/ui/CheckButton";
-import NavBar from "@/components/shared/header/NavBar";
-import Header from "@/components/shared/header/Header";
-import RecipeInfo from "@/components/shared/RecipeInfo";
+import NavBar from "@/components/layout/header/NavBar";
+import Header from "@/components/layout/header/Header";
+import RecipeInfo from "@/components/recipes/RecipeInfo";
 import CreatorInfo from "@/components/shared/CreatorInfo";
-import NavLinks from "@/components/shared/header/NavLinks";
-import LoginForm from "@/components/shared/forms/LoginForm";
+import NavLinks from "@/components/layout/header/NavLinks";
+import LoginForm from "@/components/forms/LoginForm";
 import Promosection from "@/components/shared/Promosection";
-import DirectionItem from "@/components/shared/DirectionItem";
-import DirectionList from "@/components/shared/DirectionList";
-import RecipeInfoBar from "@/components/shared/RecipeInfoBar";
-import RecipeCard from "@/components/shared/cards/RecipeCard";
-import MenuButton from "@/components/shared/header/MenuButton";
-import RecipesSection from "@/components/shared/RecipesSection";
-import IngredientList from "@/components/shared/IngredientList";
-import IngredientItem from "@/components/shared/IngredientItem";
-import RegisterForm from "@/components/shared/forms/RegisterForm";
-import CategoryCard from "@/components/shared/cards/CategoryCard";
-import CategoriesSection from "@/components/shared/CategoriesSection";
-import DirectionsSection from "@/components/shared/DirectionsSection";
-import IngredientsSection from "@/components/shared/IngredientsSection";
-import RecipeCardSkeleton from "@/components/shared/RecipeCardSkeleton";
-import FeaturedSlider from "@/components/shared/featured/FeaturedSlider";
-import FeaturedSection from "@/components/shared/featured/FeaturedSection";
-import RecipeSectionSkeleton from "@/components/shared/RecipeSectionSkeleton";
-import FeaturedRecipeCard from "@/components/shared/featured/FeaturedRecipeCard";
-import { NutritionInformationCard } from "@/components/shared/NutritionInformationCard";
-import FeaturedRecipeSkeleton from "@/components/shared/featured/FeaturedRecipeSkeleton";
+import DirectionItem from "@/components/directions/DirectionItem";
+import DirectionList from "@/components/directions/DirectionList";
+import RecipeInfoBar from "@/components/recipes/RecipeInfoBar";
+import RecipeCard from "@/components/recipes/RecipeCard";
+import MenuButton from "@/components/layout/header/MenuButton";
+import RecipesSection from "@/components/recipes/RecipesSection";
+import IngredientList from "@/components/ingredients/IngredientList";
+import IngredientItem from "@/components/ingredients/IngredientItem";
+import RegisterForm from "@/components/forms/RegisterForm";
+import CategoryCard from "@/components/categories/CategoryCard";
+import CategoriesSection from "@/components/categories/CategoriesSection";
+import DirectionsSection from "@/components/directions/DirectionsSection";
+import IngredientsSection from "@/components/ingredients/IngredientsSection";
+import RecipeCardSkeleton from "@/components/recipes/RecipeCardSkeleton";
+import FeaturedSlider from "@/components/featured/FeaturedSlider";
+import FeaturedSection from "@/components/featured/FeaturedSection";
+import RecipeSectionSkeleton from "@/components/recipes/RecipeSectionSkeleton";
+import FeaturedRecipeCard from "@/components/featured/FeaturedRecipeCard";
+import { NutritionInformationCard } from "@/components/recipes/NutritionInformationCard";
+import FeaturedRecipeSkeleton from "@/components/featured/FeaturedRecipeSkeleton";
 
 export {
     Logo,
