+++ frontend/foodieland-app/components/ingredients/IngredientItem.tsx
