+++ frontend/foodieland-app/components/ingredients/IngredientsSection.tsx
