+++ frontend/foodieland-app/components/layout/Footer.tsx
