+++ frontend/foodieland-app/components/recipes/NutritionInformationCard.tsx
